requirements.txt
influxdb==5.3.2
python-dotenv==1.0.1
tzdata
requests
matplotlib
